Input Manager working
It is able to detect mouse and keyboard events for the moment. It was tested and now window is able to fullscreen correctly. I only need to check in which screen is the window to complete all the tasks from window.

Input manager still needs gamepad to be completed. Also, note that I'm not using real time systems yet to check trigger or pressed events. That will come with time system.

Finally, I was able to add the precompiled header