Spaces successfully implemented
They work correctly. They are used to store objects in different spaces, so that we can separate in having different object spaces.

Also, fixed how objects should be removed and now it works perfectly.

Note that objects store their index on the space array and the index on their parent's children array so that it is easier to search for them. Going for object manager tomorrow.