Forgot to push editor, but I continued today
Before I continued I added imgui to the project and setup the basics of ImGui to be able to render with it.

Now I managed to create a simple scene manager and an object picker. But need to develop component editor and more things. Will do it as soon as I finish with final examns.