Clean Up on scene visualizer
Scene visualizer almost works correctly. Need to fix adding children to objects of the same parent and check that I cannot add a child to a child of the object that I want to make the child (If 0 is parent and 1 is child, not make 0 child of 1)

Apart from that, drag and drop is completed on both spaces and objects and childrens are shown correctly.