Physics added to Engine
Finally, I was able to implement at least for the moment, AABB vs AABB collisions with dynamic and static objects. It also includes the physics manager.

Both components have been added to the editor but need customization. Also added play button to editor to play the scene. Fixed serialization for that.

Render fixed since depth was not enabled. Camera also fixed and now it renders scene much clearer.

Now the problem comes in GraphicsManager since render is not working when switching to graphics manager. So now I will be working first on fixing that before adding more tests or making the rigid bodies better.