FMOD added and sound implemented
For the moment only 2D sounds are played with no positional sound, but sound is managed correctly.

Sound manager deals with channels and volumes while audio emitters deal with the sound they hold, their type of sound and their properties.

Audio listeners have been created too but their usage is only to be there so that audio can be played.